+++ app/dev/commands/app_command_test.rb
@@ -1,0 +1,7 @@
+class AppCommandTest < Liza::CommandTest
+
+  test :subject_class do
+    assert subject_class == AppCommand
+  end
+
+end