+++ app/dev/commands/macro_child_command_test.rb
@@ -1,0 +1,7 @@
+class MacroChildCommandTest < AppCommandTest
+
+  test :subject_class do
+    assert subject_class == MacroChildCommand
+  end
+
+end