+++ app/dev/commands/macro_parent_command_test.rb
@@ -1,0 +1,7 @@
+class MacroParentCommandTest < AppCommandTest
+
+  test :subject_class do
+    assert subject_class == MacroParentCommand
+  end
+
+end