+++ app/dev_box.rb
@@ -1,0 +1,13 @@
+class DevBox < Liza::DevBox
+
+  commands do
+    # Set up your command panel per the DSL in http://guides.lizarb.com/panels/command.html
+
+  end
+
+  logs do
+    # Set up your log panel per the DSL in http://guides.lizarb.com/panels/log.html
+
+  end
+
+end