+++ lib/dev_system.rb
@@ -1,0 +1,8 @@
+class DevSystem < Liza::System
+  class Error < Liza::Error; end
+
+  #
+
+  set :log_color, :green
+
+end