+++ lib/dev_system/dev/controllers/command/echo_command.rb
@@ -1,0 +1,11 @@
+class DevSystem
+  class EchoCommand < Command
+
+    def self.call args
+      log :higher, "Called #{self} with args #{args}"
+
+      raise RuntimeError, args.to_s
+    end
+
+  end
+end