+++ lib/dev_system/dev/controllers/command/echo_command_test.rb
@@ -1,0 +1,14 @@
+class DevSystem
+  class EchoCommandTest < CommandTest
+
+    test :subject_class do
+      assert subject_class == DevSystem::EchoCommand
+    end
+
+    test :settings do
+      assert subject_class.log_level == :normal
+      assert subject_class.log_color == :green
+    end
+
+  end
+end