+++ lib/dev_system/dev/controllers/command/version_command.rb
@@ -1,0 +1,9 @@
+class DevSystem
+  class VersionCommand < Command
+
+    def self.call args
+      puts Lizarb::VERSION
+    end
+
+  end
+end